Moved the calculate possible moves into the chess logic class and adjusted checkmate logic
