+++ src/test/java/ChessboardTest.java
@@ -52,12 +52,6 @@ class ChessboardTest {
     }
 
     @Test
-    void isInvalidMoveSafe() {
-        new Move(2, 0, board.getPiece(4, 7), null, board);
-        assertTrue(new ChessLogic(board).isMoveUnsafe(7, 2, board.getPiece(7, 1)));
-    }
-
-    @Test
     void queenNextToKingInCheck(){
         new Move(2, 0, board.getPiece(4, 7), null, board);
         assertTrue(new ChessLogic(board).isKingInCheck(PieceColour.WHITE));
